Update 5.database.md
this url params was not working for me. Then I went and looked at this file: https://github.com/nitrojs/nitro/blob/v2/src/runtime/internal/database.ts#L3

It seems each connector is waiting for an option object. Maybe the confusion comes from the fact that postgresql accept options to be a string as well.